+++ api/package.json
@@ -28,7 +28,6 @@
     "pg": "^8.5.1",
     "sequelize": "^6.3.5",
     "supertest-as-promised": "^4.0.2",
-    "uuid": "^8.3.2",
     "uuidv4": "^6.2.7"
   },
   "devDependencies": {
