+++ client/package.json
@@ -8,7 +8,6 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
     "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
     "@sweetalert/with-react": "^0.1.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
@@ -19,7 +18,6 @@
     "mercadopago": "^1.5.7",
     "mkcert": "^1.4.0",
     "react": "^16.13.1",
-    "react-bootstrap": "^1.5.2",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.3",
     "react-router-dom": "^5.2.0",
@@ -28,7 +26,6 @@
     "react-speech-recognition": "^3.8.0",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0",
-    "sweetalert": "^2.1.2",
     "sweetalert2": "^10.16.3",
     "universal-cookie": "^4.0.4"
   },
