+++ client/src/components/AddCategory.jsx
@@ -6,6 +6,7 @@ import {
   addCategory,
   deleteSuccess,
 } from "../redux/actions/categoryActions";
+
 import {
   Grid,
   Container,
@@ -14,8 +15,9 @@ import {
   Box,
   TextField,
 } from "@material-ui/core";
+
 import { sweetAlert } from "../helpers/utils";
-import makeStyles from "./AddCategorySytles";
+import makeStyles from "./componentsStyles/AddCategorySytles";
 
 const initialState = {
   name: "",
