+++ client/src/components/SearchBar.jsx
@@ -3,8 +3,7 @@ import { useDispatch } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { searchProducts } from "../redux/actions/productActions";
 //----------- ↓ Import Styles ↓ -----------
-import { makeStyles } from "@material-ui/core/styles";
-import { form, IconButton, Input, InputAdornment } from "@material-ui/core";
+import { IconButton, Input, InputAdornment, makeStyles } from "@material-ui/core";
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
